more consolidation, removing gfycat random gifs for behrad/keanu
