+++ README.md
@@ -11,6 +11,4 @@ KEANU_BOT_TOKEN
 SPRING_DATASOURCE_USERNAME
 SPRING_DATASOURCE_PASSWORD
 SPRING_DATASOURCE_URL
-GFYCAT_CLIENT_ID
-GFYCAT_CLIENT_SECRET
 ```
