+++ src/main/java/com/badfic/philbot/config/BaseConfig.java
@@ -1,8 +1,28 @@
 package com.badfic.philbot.config;
 
+import static net.dv8tion.jda.api.requests.GatewayIntent.DIRECT_MESSAGES;
+import static net.dv8tion.jda.api.requests.GatewayIntent.GUILD_BANS;
+import static net.dv8tion.jda.api.requests.GatewayIntent.GUILD_MEMBERS;
+import static net.dv8tion.jda.api.requests.GatewayIntent.GUILD_MESSAGES;
+import static net.dv8tion.jda.api.requests.GatewayIntent.GUILD_MESSAGE_REACTIONS;
+import static net.dv8tion.jda.api.requests.GatewayIntent.GUILD_VOICE_STATES;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
+import com.jagrosh.jdautilities.command.Command;
+import com.jagrosh.jdautilities.command.CommandClient;
+import com.jagrosh.jdautilities.command.CommandClientBuilder;
+import com.neovisionaries.ws.client.WebSocketFactory;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
+import net.dv8tion.jda.api.JDA;
+import net.dv8tion.jda.api.JDABuilder;
+import net.dv8tion.jda.api.entities.Activity;
+import net.dv8tion.jda.api.hooks.EventListener;
+import net.dv8tion.jda.internal.utils.IOUtil;
+import okhttp3.OkHttpClient;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,6 +31,53 @@ import org.springframework.core.io.ClassPathResource;
 
 @Configuration
 public class BaseConfig {
+
+    private static final String[] WATCHING_STATUS_LIST = {
+            "Bill & Ted Face the Music",
+            "The SpongeBob Movie: Sponge on the Run",
+            "Toy Story 4",
+            "Always Be My Maybe",
+            "John Wick: Chapter 3 - Parabellum",
+            "Replicas",
+            "Destination Wedding",
+            "Siberia",
+            "John Wick: Chapter 2",
+            "Keanu",
+            "John Wick",
+            "Constantine",
+            "Something's Gotta Give",
+            "The Matrix Revolutions",
+            "The Matrix Reloaded",
+            "The Matrix",
+            "Speed",
+            "Bill & Ted's Bogus Journey",
+            "Bill & Ted's Excellent Adventure"
+    };
+
+    private static final String[] PLAYING_STATUS_LIST = {
+            "Tekken",
+            "in Heyworld",
+            "Street Fighter",
+            "Overwatch",
+            "Super Smash Bros",
+            "Dragon Age Origins",
+            "Ocarina of Time",
+            "Mass Effect",
+            "Mario & Sonic at the Olympic Games",
+            "Sonic the Hedgehog",
+            "Super Mario Bros",
+            "Resident Evil",
+            "Wii Sports"
+    };
+
+    @Value("${PHIL_BOT_TOKEN}")
+    public String philBotToken;
+
+    @Value("${KEANU_BOT_TOKEN}")
+    public String keanuBotToken;
+
+    @Value("${BEHRAD_BOT_TOKEN}")
+    public String behradBotToken;
 
     @Value("${git.commit.id}")
     public String commitSha;
@@ -21,20 +88,9 @@ public class BaseConfig {
     @Value("${OWNER_ID}")
     public String ownerId;
 
-    @Value("${GFYCAT_CLIENT_ID:}")
-    public String gfycatClientId;
-
-    @Value("${GFYCAT_CLIENT_SECRET:}")
-    public String gfycatClientSecret;
-
     @Bean
     public ObjectMapper objectMapper() {
         return new ObjectMapper();
-    }
-
-    @Bean(name = "gfycatClient", destroyMethod = "close")
-    public CloseableHttpClient gfycatClient() {
-        return HttpClients.createDefault();
     }
 
     @Bean
@@ -45,6 +101,60 @@ public class BaseConfig {
         propsConfig.setIgnoreResourceNotFound(true);
         propsConfig.setIgnoreUnresolvablePlaceholders(true);
         return propsConfig;
+    }
+
+    @Bean
+    public OkHttpClient okHttpClient() {
+        return IOUtil.newHttpClientBuilder().build();
+    }
+
+    @Bean
+    public WebSocketFactory webSocketFactory() {
+        return new WebSocketFactory();
+    }
+
+    @Bean(name = "behradJda")
+    public JDA behradJda(OkHttpClient okHttpClient, WebSocketFactory webSocketFactory) throws Exception {
+        return JDABuilder.createLight(behradBotToken, Collections.emptyList())
+                .setHttpClient(okHttpClient)
+                .setWebsocketFactory(webSocketFactory)
+                .setActivity(Activity.playing(PLAYING_STATUS_LIST[ThreadLocalRandom.current().nextInt(PLAYING_STATUS_LIST.length)]))
+                .build();
+    }
+
+    @Bean(name = "keanuJda")
+    public JDA keanuJda(OkHttpClient okHttpClient, WebSocketFactory webSocketFactory) throws Exception {
+        return JDABuilder.createLight(keanuBotToken, Collections.emptyList())
+                .setHttpClient(okHttpClient)
+                .setWebsocketFactory(webSocketFactory)
+                .setActivity(Activity.watching(WATCHING_STATUS_LIST[ThreadLocalRandom.current().nextInt(WATCHING_STATUS_LIST.length)]))
+                .build();
+    }
+
+    @Bean(name = "philCommandClient")
+    public CommandClient philCommandClient(List<Command> commands) {
+        return new CommandClientBuilder()
+                .setOwnerId(ownerId)
+                .setPrefix("!!")
+                .useHelpBuilder(false)
+                .addCommands(commands.stream().filter(c -> c instanceof PhilMarker).toArray(Command[]::new))
+                .setActivity(Activity.playing("with our feelings"))
+                .setEmojis("\uD83C\uDF83", "⚠️", "\uD83D\uDC80")
+                .build();
+    }
+
+    @Bean(name = "philJda")
+    public JDA philJda(OkHttpClient okHttpClient,
+                       WebSocketFactory webSocketFactory,
+                       List<EventListener> eventListeners,
+                       @Qualifier("philCommandClient") CommandClient philCommandClient) throws Exception {
+        return JDABuilder.create(philBotToken, Arrays.asList(GUILD_MEMBERS, GUILD_BANS, GUILD_MESSAGES, GUILD_VOICE_STATES, GUILD_MESSAGE_REACTIONS, DIRECT_MESSAGES))
+                .setHttpClient(okHttpClient)
+                .setWebsocketFactory(webSocketFactory)
+                .addEventListeners(eventListeners.stream().filter(e -> e instanceof PhilMarker).toArray(EventListener[]::new))
+                .addEventListeners(philCommandClient)
+                .setActivity(Activity.playing("with our feelings"))
+                .build();
     }
 
 }
